requirements.txt
arrow==0.14.7
asgiref==3.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
convertdate==2.2.2
Django==3.1.3
django-rest-framework==0.1.0
djangorestframework==3.12.1
holidays==0.10.3
ics==0.7
idna==2.10
korean-lunar-calendar==0.2.1
psycopg2-binary==2.8.6
PyMeeus==0.3.7
python-dateutil==2.8.1
pytz==2020.4
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tqdm==4.51.0
urllib3==1.25.11
